PLT-586 - Creation of an immutable FxMatrix class
This PR contains a new immutable FX Matrix for
currency conversions. Whilst the data structure
is more or less the same as the original FxMatrix
in v2, there is no mutability. Instead a mutable
builder is provided.

The other main differences to the original class
are:
 - the ability to add a rate with no currencies in
   common with the matrix with the expectation that
   other additions will allow it to be added
 - the ability to handle a collection of rates
 - the ability to collect a stream of rates directly
   into a new FxMatrix

The builder has been implemented such that its
underlying matrix is not resized for every currency
addition, but only periodically (it doubles in size
when needed). This hopefully balances between too much
resizing and too much memory but interested in other
views.

Test coverage is at 100%.