PLT-600 - new time series
This adds a new time series implementation which stores
its data in a more efficient way (as long as the data
is dense). This is primarily to allow whole timseries
to be saved and restored from database masters as a
blob.

The builder is capable of determining which implementation
is more appropriate based on the data.

There are a few of items it would be useful to get
additional feedback on:

- the dense implementation has of() methods matching
  the original sparse version. It is questionable
  whether these are needed (as for the sparse case
  also) and perhaps users should be nudged towards the
  builder.
- the dense implementation asks the client to specify
  whether weekends are stored. However, as the data
  implicitly contains this information the client
  shouldn't need to specify. The builder already does
  this determination so is another reason to remove the
  of methods.
- the dense implementation's of methods all do
  similar validations. If they are kept then this
  should be moved to a single internal analysis class
  that can check for nulls, determine date extremes,
  and determine weekends