Demonstrate issues with new formatter
The formatter config is included so all settings can be replicated.

The code in these two projects (strata-basics and strata-collect)
was fully acceptable to the old formatter.
ie. using Right-click/Source/Format on the project resulted in no change
The rest of this commit is the result of running the new formatter
in the same way.

As can be seen, the new formatter results in dramatically different output.