Support Fixed Initial Notional on MTM leg of XCCY swaps.
The fixed initial notional is specified at the FxResetCalculation level, as it is only applicable for legs with Fx Reset.
When resolving payment schedule the fixed notional amount is used for the first period (jf present)

FxResetCalculation.resolve() now returns an Optional FxReset, to account for the fact that the fx reset may not be present for the initial leg. The index of the period is also passed in order to ensure that the override is only applied to the first leg.

This changes models the FpML concept of initialValue on fxLinkedNotionalSchedule:
http://www.fpml.org/spec/fpml-5-6-1-wd-1/html/pretrade/schemaDocumentation/schemas/fpml-ird-5-6_xsd/complexTypes/Calculation/fxLinkedNotionalSchedule.html