Use actual start month and end month rather than proleptic months to determine EOM rolls.
This fixes an edge case for February. For EOM trades, If start and end date are in February and one (but not both) was in a a leap year then the roll convention was calculated as 29 rather than EOM.

In terms of existing behaviour, only other change is to the below use case:
* If start and end date are in same month but not the same year
* With different day of month
* End date day of month is the last day of the month

This case previously returned EOM roll day but now returns start day of month as the roll day. A trade with this combination of regular period start/end date should only be valid if it is a single period swap, in which case roll day would not be relevant.