Ensure only one direction per FX pair is considered conventional
Prior to this there could be a situation where both directions would return true for isConventional() calls.

This was true for all USD/metal pairs (e.g. USD/XAU) and occurred because one direction was listed in the static file (XAU/USD) but the first currency (XAU) had a lower priority order than the second (USD).

Thus  isConventional returned true for USD/XAU because USD is higher priority and XAU/USD also returned true because that pair is in the file