Fix Hungarian working Saturdays for 2020.
Per below sources, the August working Saturday for 2020 is 1 week after the long weekend.
https://helpers.hu/other/bank-holidays-in-hungary-in-2020/
https://www.timeanddate.com/holidays/hungary/2020

The last time that National Day fell on a Thursday was 2015. On this occasion the working Saturday was two weeks prior to the long weekend. Assumedly that's where we derived our logic from

So unclear what the logic is for deciding where the working Saturday falls, it may not even be algorithmic.

This change ensures we get the right date for 2020.